Add email verification to submissions
For each submission there will be an email sent to the user@binghamton.edu.
The email will include a hash of the username_assignment. This can be used
to check that the user is being truthful (just ask for the hash). This is of
course not very secure - if the user knows how we create the verification
codes then we are thrwarted.

If we combine this method with a uniquely incrementing identifier then
we are in a lot better business.

Reference #1